fix(otel): repair stale livetests
Two livetests were broken at the repo level but only get exercised
under the livetest job in CI, so the breakage didn't surface until
the v1.0.0 CI run.

- transports/otellog/livetest_test.go: TestLive_StructMetadataNestedThroughSDK
  asserted .AsFloat64()==7 on the int field of a struct, with a
  comment explaining the JSON roundtrip turned ints into float64.
  After the structured-transport rewrite (commit adb7114) struct
  walking is reflective and preserves the runtime type, so the field
  arrives as KindInt64. Switch the assertion to .AsInt64() and update
  the comment.

- plugins/oteltrace/livetest_test.go: newLiveSetup called a
  non-existent setup() helper. Replace with plugintest.Install (the
  helper used by every other oteltrace test), import plugintest.

Both livetests are run by the .github/workflows/ci.yml Livetest job
under -tags=livetest. Now green locally.

+++ transports/otellog/livetest_test.go
@@ -187,8 +187,9 @@ func TestLive_StructMetadataNestedThroughSDK(t *testing.T) {
 		t.Fatalf("metadata attr should be Map, got %v", meta.Kind())
 	}
 	nested := kvMap(meta.AsMap())
-	// JSON roundtrip turns Go ints into float64.
-	if nested["id"].AsFloat64() != 7 || nested["name"].AsString() != "Alice" {
+	// MetadataAsMap walks the struct via reflection and preserves the
+	// runtime type, so an int field arrives as KindInt64.
+	if nested["id"].AsInt64() != 7 || nested["name"].AsString() != "Alice" {
 		t.Errorf("metadata fields: got %v", nested)
 	}
 }
